% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymptoticMeanH_JSDivMultinomial.R
\name{asymptoticMeanH_JSDivMultinomial}
\alias{asymptoticMeanH_JSDivMultinomial}
\title{Function that computes the asymptotic mean of the sum and difference
between the normalized Shannon entropy and the Jensen-Shannon divergence
under the Multinomial model}
\usage{
asymptoticMeanH_JSDivMultinomial(p, operation)
}
\arguments{
\item{p}{a probability function}

\item{operation}{"s" for the sum and "d" for de difference}
}
\value{
A scalar
}
\description{
Function that computes the asymptotic mean of the sum and difference
between the normalized Shannon entropy and the Jensen-Shannon divergence
under the Multinomial model
}
\keyword{internal}
