% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymptoticMeanJSDivergenceMultinomial.R
\name{asymptoticMeanJSDivergenceMultinomial}
\alias{asymptoticMeanJSDivergenceMultinomial}
\title{Function that computes the asymptotic mean of the Jensen-Shannon divergence
under the Multinomial model}
\usage{
asymptoticMeanJSDivergenceMultinomial(p)
}
\arguments{
\item{p}{a probability function}
}
\value{
The asymptotic mean (a scalar)
}
\description{
Function that computes the asymptotic mean of the Jensen-Shannon divergence
under the Multinomial model
}
\keyword{internal}
