% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymptoticVarH_JSDivMultinomial.R
\name{asymptoticVarH_JSDivMultinomial}
\alias{asymptoticVarH_JSDivMultinomial}
\title{Function that computes the asymptotic variance of the sum and difference
between the normalized Shannon entropy and the Jensen-Shannon divergence
under the Multinomial model}
\usage{
asymptoticVarH_JSDivMultinomial(p, n, operation)
}
\arguments{
\item{p}{a probability function}

\item{n}{the number of independent trials}

\item{operation}{"s" for the sum and "d" for de difference}
}
\value{
A scalar
}
\description{
Function that computes the asymptotic variance of the sum and difference
between the normalized Shannon entropy and the Jensen-Shannon divergence
under the Multinomial model
}
\keyword{internal}
