% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymptoticVarHShannonMultinomial.R
\name{asymptoticVarHShannonMultinomial}
\alias{asymptoticVarHShannonMultinomial}
\title{Function that computes the asymptotic variance of the normalized Shannon
entropy under the Multinomial model}
\usage{
asymptoticVarHShannonMultinomial(p, n)
}
\arguments{
\item{p}{a probability function}

\item{n}{the number of independent trials}
}
\value{
A scalar
}
\description{
Function that computes the asymptotic variance of the normalized Shannon
entropy under the Multinomial model
}
