% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymptoticVarJSDivergenceMultinomial.R
\name{asymptoticVarJSDivergenceMultinomial}
\alias{asymptoticVarJSDivergenceMultinomial}
\title{Function that computes the asymptotic variance of the Jensen-Shannon divergence
under the Multinomial model}
\usage{
asymptoticVarJSDivergenceMultinomial(p, n)
}
\arguments{
\item{p}{a probability function}

\item{n}{the number of independent trials}
}
\value{
A scalar
}
\description{
Function that computes the asymptotic variance of the Jensen-Shannon divergence
under the Multinomial model
}
\keyword{internal}
