% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationH_JSDivMultinomial.R
\name{correlationH_JSDivMultinomial}
\alias{correlationH_JSDivMultinomial}
\title{Function that computes the correlation coefficient
between the normalized Shannon entropy and the Jensen-Shannon divergence
under the Multinomial model}
\usage{
correlationH_JSDivMultinomial(p, n)
}
\arguments{
\item{p}{a probability function}

\item{n}{the number of independent trials}
}
\value{
A scalar
}
\description{
Function that computes the correlation coefficient
between the normalized Shannon entropy and the Jensen-Shannon divergence
under the Multinomial model
}
\keyword{internal}
