% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_pos.R
\name{ind_pos}
\alias{ind_pos}
\title{Function that assigns the corresponding OP to a partition (word)}
\usage{
ind_pos(part)
}
\arguments{
\item{part}{a partition of a time series}
}
\description{
Internal function. Candidate for C++ implementation.
}
