% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinfLsup.R
\docType{data}
\name{LinfLsup}
\alias{LinfLsup}
\title{Boundaries of the \eqn{H\times C} closed manifold}
\format{
\subsection{\code{LinfLsup}}{

A data frame with 184936 rows and 4 variables:
\describe{
\item{H}{entropy (value)}
\item{C}{statistical complexity (value)}
\item{Dimension}{embedding dimension (factor, "3", "4", "5", "6")}
\item{Side}{Lower or Upper side of the feasible region (factor, "Lower" and "Upper")}
}
}
}
\source{
\url{https://dx.doi.org/10.1016/j.physa.2005.11.053}
LinfLsup
}
\usage{
LinfLsup
}
\description{
A data frame with the lower and upper boundaries of the \eqn{H\times C} closed manifold
for \eqn{D=3,4,5,6}.
}
\keyword{datasets}
