% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPseq.R
\name{OPseq}
\alias{OPseq}
\title{Function that finds the OP sequence of a time series, given the
embedding dimension and lag}
\usage{
OPseq(TS, emb, lag = 1)
}
\arguments{
\item{TS}{time series}

\item{emb}{embedding dimension}

\item{lag}{time lag (default value: 1)}
}
\description{
Function that finds the OP sequence of a time series, given the
embedding dimension and lag
}
\examples{
# Generate a time series and compute its ordinal patterns
set.seed(1234567890, kind="Mersenne-Twister")
x <- rnorm(1000) # white noise
OPseq(x, emb=4, lag=1)

}
