% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm.R
\name{perm}
\alias{perm}
\title{Function that finds all possible ordinal patterns of the
elements \eqn{0, 1, ... , m-1}}
\usage{
perm(v)
}
\arguments{
\item{v}{a vector \eqn{0, 1, \dots , m-1}}
}
\description{
Function that finds all possible ordinal patterns of the
elements \eqn{0, 1, ... , m-1}
}
\keyword{internal}
