% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_i.R
\name{pi_i}
\alias{pi_i}
\title{Function that assigns the label i to an}
\usage{
pi_i(pat)
}
\arguments{
\item{pat}{an OP, for example, pat = 0,1,2 if \eqn{x_1 < x_2 < x_3}}
}
\description{
Function that assigns the label i to an
}
