% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qmatrix.R
\name{Qmatrix}
\alias{Qmatrix}
\title{Function that computes the Ordinal Patterns transitions matrices}
\usage{
Qmatrix(TS, emb, lag)
}
\arguments{
\item{TS}{time series}

\item{emb}{embedding dimension}

\item{lag}{the lag}
}
\value{
an \eqn{m!\times m!} matrix of probabilities
}
\description{
Let \eqn{\bm \pi=(\pi_1,\pi_2,\dots,\pi_n)} be the sequence of ordinal patterns
from the time series \eqn{\bm x=(x_1,x_2,\dots,x_{n-m+1})} computed over windows
of embedding dimension \eqn{m}. For each lag \eqn{\ell =1,2, \dots , m-1}, the ordinal
pattern transition matrix of order \eqn{\ell} is
\eqn{\textbf{Q}^{(\ell)} \in \mathbb{R}^{m! \times m!}}, with elements
\deqn{q_{ij}^{(\ell)} = \Pr\big(\pi_t=\pi^{(i)} \wedge \pi_{t+\ell}=\pi^{(j)}) = q_i \Pr(\pi_{t+\ell} = \pi^{(j)} \mid \pi_t=\pi^{(i)}\big),}
for \eqn{i,j=1,2,\dots ,m!}.
}
