% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sigma.R
\name{Sigma}
\alias{Sigma}
\title{Function that computes the covariance matrix of ordinal patterns
from a time series, given embedding dimension}
\usage{
Sigma(TS, emb)
}
\arguments{
\item{TS}{time series}

\item{emb}{embedding dimension}
}
\value{
A covariance matrix
}
\description{
Function that computes the covariance matrix of ordinal patterns
from a time series, given embedding dimension
}
\keyword{internal}
