% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SigmaMultinomial.R
\name{SigmaMultinomial}
\alias{SigmaMultinomial}
\title{Function that computes the covariance matrix of the Multinomial model}
\usage{
SigmaMultinomial(p)
}
\arguments{
\item{p}{a probability function}
}
\value{
A covariance matrix
}
\description{
Function that computes the covariance matrix of the Multinomial model
}
\keyword{internal}
