% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sigmaq.R
\name{Sigmaq}
\alias{Sigmaq}
\title{Function that computes the covariance matrix for a time series}
\arguments{
\item{TS}{time series}

\item{emb}{embedding dimension}

\item{ent}{type of entropy: S (Shannon), T (Tsallis, \eqn{beta} is required), R (Rényi, \eqn{beta} is required), or F (Fisher)}

\item{beta}{parameter for the Tsallis and Rényi entropies}
}
\description{
Function that computes the covariance matrix for a time series
}
\keyword{internal}
