% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComplexity.R
\name{StatComplexity}
\alias{StatComplexity}
\title{Statistical Complexity of a probability function}
\usage{
StatComplexity(p)
}
\arguments{
\item{p}{a probability function}
}
\value{
a value in \eqn{[0,1]}
}
\description{
Given a probability function \eqn{\bm p}, its statistical complexity
is the product of its entropy \eqn{H(\bm p)} and its distance to an
equilibrium distribution \eqn{\bm u}.
}
\details{
The normalized Jensen-Shannon distance to the uniform distribution
\eqn{\bm u=(1/k, 1/k,\dots,1/k)}, where \eqn{k\geq 2} is the size of \eqn{p},
is \eqn{Q=Q'/\max{Q'}}, where
\deqn{Q'(\bm{p}, \bm{u}) = \sum_{\ell=1}^{k} \Big[\big(p(\ell)-\frac{1}{k}\big) \ln\big(k p(\ell)\big)\Big],}
and
\deqn{\max Q' = -2\Big[\frac{k!+1}{k!} \ln(k!+1) - 2 \ln(2k!) + \ln k!\Big].}
}
\examples{

# Time series
set.seed(1234567890, kind="Mersenne-Twister")
x <- rnorm(1000) # white noise
y <- mov.av(x, order=50) # smoothed with moving averages
# Statistical Complexity of patterns of dimension 4
StatComplexity(OPprob(x, emb=4))
StatComplexity(OPprob(y, emb=4))

}
