% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varC.R
\name{varC}
\alias{varC}
\title{Function that computes the variance of the Normal distribution that approximates
the asymptotic probability density function of the
statistical complexity under the Multinomial model}
\usage{
varC(p, n)
}
\arguments{
\item{p}{a probability function}

\item{n}{the number of independent trials}
}
\value{
A scalar
}
\description{
Function that computes the variance of the Normal distribution that approximates
the asymptotic probability density function of the
statistical complexity under the Multinomial model
}
